refactor: change getGoogleUrl name
